+++ lib/jsonapi-resources.rb
@@ -41,8 +41,8 @@ require 'jsonapi/operation_result'
 require 'jsonapi/callbacks'
 require 'jsonapi/link_builder'
 require 'jsonapi/active_relation/adapters/join_left_active_record_adapter'
-require 'jsonapi/active_relation/join_manager'
-require 'jsonapi/active_relation/join_manager_v10'
+require 'jsonapi/active_relation/join_manager_through_inverse'
+require 'jsonapi/active_relation/join_manager_through_primary'
 require 'jsonapi/resource_identity'
 require 'jsonapi/resource_fragment'
 require 'jsonapi/resource_tree'
