+++ test/unit/active_relation_resource_finder/join_manager_through_primary_test.rb
@@ -1,9 +1,9 @@
 require File.expand_path('../../../test_helper', __FILE__)
 require 'jsonapi-resources'
 
-class JoinManagerV10Test < ActiveSupport::TestCase
+class JoinManagerThroughPrimaryTest < ActiveSupport::TestCase
   def test_no_added_joins
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource)
 
     records = PostResource.records({})
     records = join_manager.join(records, {})
@@ -14,7 +14,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
   def test_add_single_join
     filters = {'tags' => ['1']}
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource, filters: filters)
     records = PostResource.records({})
     records = join_manager.join(records, {})
     assert_equal 'SELECT "posts".* FROM "posts" LEFT OUTER JOIN "posts_tags" ON "posts_tags"."post_id" = "posts"."id" LEFT OUTER JOIN "tags" ON "tags"."id" = "posts_tags"."tag_id"', sql_for_compare(records.to_sql)
@@ -24,7 +24,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
   def test_joins_have_join_options
     filters = {'tags' => ['1']}
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource, filters: filters)
     records = PostResource.records({})
     records = join_manager.join(records, {})
     assert_equal 'SELECT "posts".* FROM "posts" LEFT OUTER JOIN "posts_tags" ON "posts_tags"."post_id" = "posts"."id" LEFT OUTER JOIN "tags" ON "tags"."id" = "posts_tags"."tag_id"', sql_for_compare(records.to_sql)
@@ -38,7 +38,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
   def test_add_single_sort_join
     sort_criteria = [{field: 'tags.name', direction: :desc}]
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource, sort_criteria: sort_criteria)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource, sort_criteria: sort_criteria)
     records = PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -50,7 +50,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
   def test_add_single_sort_and_filter_join
     filters = {'tags' => ['1']}
     sort_criteria = [{field: 'tags.name', direction: :desc}]
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource, sort_criteria: sort_criteria, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource, sort_criteria: sort_criteria, filters: filters)
     records = PostResource.records({})
     records = join_manager.join(records, {})
     assert_equal 'SELECT "posts".* FROM "posts" LEFT OUTER JOIN "posts_tags" ON "posts_tags"."post_id" = "posts"."id" LEFT OUTER JOIN "tags" ON "tags"."id" = "posts_tags"."tag_id"', sql_for_compare(records.to_sql)
@@ -64,7 +64,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
       'author' => ['1']
     }
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource, filters: filters)
     records = PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -76,8 +76,8 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
 
   def test_add_joins_source_relationship
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource,
-                                                            source_relationship: PostResource._relationship(:comments))
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource,
+                                                                          source_relationship: PostResource._relationship(:comments))
     records = PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -87,8 +87,8 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
 
   def test_add_joins_source_relationship_with_custom_apply
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: Api::V10::PostResource,
-                                                            source_relationship: Api::V10::PostResource._relationship(:comments))
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: Api::V10::PostResource,
+                                                                          source_relationship: Api::V10::PostResource._relationship(:comments))
     records = Api::V10::PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -106,7 +106,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
       'author' => ['1']
     }
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: Api::V10::PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: Api::V10::PostResource, filters: filters)
     records = Api::V10::PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -123,7 +123,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
       'comments.tags' => ['1']
     }
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: Api::V10::PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: Api::V10::PostResource, filters: filters)
     records = Api::V10::PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -141,7 +141,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
       'author.foo' => ['1']
     }
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: Api::V10::PostResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: Api::V10::PostResource, filters: filters)
     records = Api::V10::PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -155,8 +155,8 @@ class JoinManagerV10Test < ActiveSupport::TestCase
   def test_add_joins_with_sub_relationship
     relationships = %w(author author.comments tags)
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: Api::V10::PostResource, relationships: relationships,
-                                                            source_relationship: Api::V10::PostResource._relationship(:comments))
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: Api::V10::PostResource, relationships: relationships,
+                                                                          source_relationship: Api::V10::PostResource._relationship(:comments))
     records = Api::V10::PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -174,10 +174,10 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
     relationships = %w(author author.comments tags)
 
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PostResource,
-                                                            filters: filters,
-                                                            relationships: relationships,
-                                                            source_relationship: PostResource._relationship(:comments))
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PostResource,
+                                                                          filters: filters,
+                                                                          relationships: relationships,
+                                                                          source_relationship: PostResource._relationship(:comments))
     records = PostResource.records({})
     records = join_manager.join(records, {})
 
@@ -189,8 +189,8 @@ class JoinManagerV10Test < ActiveSupport::TestCase
   end
 
   def test_polymorphic_join_belongs_to_just_source
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PictureResource,
-                                                            source_relationship: PictureResource._relationship(:imageable))
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PictureResource,
+                                                                          source_relationship: PictureResource._relationship(:imageable))
 
     records = PictureResource.records({})
     records = join_manager.join(records, {})
@@ -204,7 +204,7 @@ class JoinManagerV10Test < ActiveSupport::TestCase
 
   def test_polymorphic_join_belongs_to_filter
     filters = {'imageable' => ['Foo']}
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PictureResource, filters: filters)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PictureResource, filters: filters)
 
     records = PictureResource.records({})
     records = join_manager.join(records, {})
@@ -221,9 +221,9 @@ class JoinManagerV10Test < ActiveSupport::TestCase
     }
 
     relationships = %w(imageable file_properties)
-    join_manager = JSONAPI::ActiveRelation::JoinManagerV10.new(resource_klass: PictureResource,
-                                                            filters: filters,
-                                                            relationships: relationships)
+    join_manager = JSONAPI::ActiveRelation::JoinManagerThroughPrimary.new(resource_klass: PictureResource,
+                                                                          filters: filters,
+                                                                          relationships: relationships)
 
     records = PictureResource.records({})
     records = join_manager.join(records, {})
